refactor start flow
